# sql/sqlbase: micro optimizations for RowContainer.At