Rework `RowContainer.getChunkAndPos` to use shifting instead of
division. Rework how `numCols==0` is handled to remove the need for
various special cases in `RowContainer.At`.

```
name              old time/op  new time/op  delta
RowContainerAt-8  5.84ns ± 2%  2.45ns ± 3%  -58.12%  (p=0.000 n=9+10)
```

Release note (performance improvement): Improve performance of low-level
row manipulation routines.