@petermattis 2017-12-13
When looking at some of the new DistSQL benchmarks, `RowContainer.At` was showing up near the top of profiles.