@RaduBerinde 2017-12-13
<img class="emoji" title=":lgtm_strong:" alt=":lgtm_strong:" align="absmiddle" src="https://reviewable.io/lgtm_strong.png" height="20" width="61"/> 

---

Review status: 0 of 2 files reviewed at latest revision, 2 unresolved discussions, some commit checks pending.

---

*[pkg/sql/sqlbase/row_container.go, line 47 at r1](https://reviewable.io:443/reviews/cockroachdb/cockroach/20688#-L0GImSS_dDS2VXYpxaa:-L0GImSS_dDS2VXYpxab:b-2lxbpc) ([raw file](https://github.com/cockroachdb/cockroach/blob/04aaf29f84448c6e301f872cfa8e17eddc2cc66d/pkg/sql/sqlbase/row_container.go#L47)):*
> ```Go
> 
> 	// rowsPerChunk is the number of rows in a chunk; we pack multiple rows in a
> 	// single []Datum to reduce the overhead of the slice if we have few columns.
> ```

mention that it must be a power of 2

---

*[pkg/sql/sqlbase/row_container_test.go, line 135 at r1](https://reviewable.io:443/reviews/cockroachdb/cockroach/20688#-L0GJ9rryzRpeF8BrJep:-L0GJ9rryzRpeF8BrJeq:b-8hahrj) ([raw file](https://github.com/cockroachdb/cockroach/blob/04aaf29f84448c6e301f872cfa8e17eddc2cc66d/pkg/sql/sqlbase/row_container_test.go#L135)):*
> ```Go
> }
> 
> var globalFalse bool
> ```

[nit] rename this to debugRowContainerAt or similar. Or remove it, it's only one line to add back if we need to debug.

---


*Comments from [Reviewable](https://reviewable.io:443/reviews/cockroachdb/cockroach/20688#-:-L0GJYuFTxtherU_to6b:b-uwkowd)*
<!-- Sent from Reviewable.io -->