@petermattis 2017-12-13
Review status: 0 of 2 files reviewed at latest revision, 2 unresolved discussions.

---

*[pkg/sql/sqlbase/row_container.go, line 47 at r1](https://reviewable.io:443/reviews/cockroachdb/cockroach/20688#-L0GImSS_dDS2VXYpxaa:-L0GQvxLEMg3dyEW18mv:b-896fix) ([raw file](https://github.com/cockroachdb/cockroach/blob/04aaf29f84448c6e301f872cfa8e17eddc2cc66d/pkg/sql/sqlbase/row_container.go#L47)):*
<details><summary><i>Previously, RaduBerinde wrote…</i></summary><blockquote>

mention that it must be a power of 2
</blockquote></details>

Done.

---

*[pkg/sql/sqlbase/row_container_test.go, line 135 at r1](https://reviewable.io:443/reviews/cockroachdb/cockroach/20688#-L0GJ9rryzRpeF8BrJep:-L0GQxQJ_Ku7_1okEC2O:bss4a2b) ([raw file](https://github.com/cockroachdb/cockroach/blob/04aaf29f84448c6e301f872cfa8e17eddc2cc66d/pkg/sql/sqlbase/row_container_test.go#L135)):*
<details><summary><i>Previously, RaduBerinde wrote…</i></summary><blockquote>

[nit] rename this to debugRowContainerAt or similar. Or remove it, it's only one line to add back if we need to debug.
</blockquote></details>

Removed.

---


*Comments from [Reviewable](https://reviewable.io:443/reviews/cockroachdb/cockroach/20688)*
<!-- Sent from Reviewable.io -->